chore: remove .idea project and bump version to 1.0.4
Removed the `.idea` project from the solution file, including its
associated `ProjectSection` and related files. Updated the `Global`
section in the solution file to remove the `.idea` project GUID.

Updated the version badge in `README.md` and the `Version` property
in `Myrtus.Clarity.Generator.Presentation.csproj` from 1.0.3 to 1.0.4
to reflect the new release.